requirements.txt
requests
pillow
Discohook